+++ src/main/kotlin/io/openfuture/chain/core/model/entity/transaction/confirmed/Transaction.kt
@@ -2,6 +2,7 @@ package io.openfuture.chain.core.model.entity.transaction.confirmed
 
 import io.openfuture.chain.core.model.entity.block.MainBlock
 import io.openfuture.chain.core.model.entity.transaction.BaseTransaction
+import io.openfuture.chain.core.model.entity.transaction.TransactionFooter
 import io.openfuture.chain.core.model.entity.transaction.TransactionHeader
 import io.openfuture.chain.network.message.core.TransactionMessage
 import javax.persistence.*
@@ -11,15 +12,13 @@ import javax.persistence.*
 @Inheritance(strategy = InheritanceType.JOINED)
 abstract class Transaction(
     header: TransactionHeader,
-    hash: String,
-    senderSignature: String,
-    senderPublicKey: String,
+    footer: TransactionFooter,
 
     @ManyToOne
     @JoinColumn(name = "block_id", nullable = false)
     var block: MainBlock
 
-) : BaseTransaction(header, hash, senderSignature, senderPublicKey) {
+) : BaseTransaction(header, footer) {
 
     abstract fun toMessage(): TransactionMessage