+++ src/main/kotlin/io/openfuture/chain/core/model/entity/transaction/unconfirmed/UnconfirmedTransferTransaction.kt
@@ -1,7 +1,7 @@
 package io.openfuture.chain.core.model.entity.transaction.unconfirmed
 
+import io.openfuture.chain.core.model.entity.transaction.TransactionFooter
 import io.openfuture.chain.core.model.entity.transaction.TransactionHeader
-import io.openfuture.chain.core.model.entity.transaction.payload.TransactionPayload
 import io.openfuture.chain.core.model.entity.transaction.payload.TransferTransactionPayload
 import io.openfuture.chain.network.message.core.TransferTransactionMessage
 import io.openfuture.chain.rpc.domain.transaction.request.TransferTransactionRequest
@@ -13,42 +13,34 @@ import javax.persistence.Table
 @Table(name = "u_transfer_transactions")
 class UnconfirmedTransferTransaction(
     header: TransactionHeader,
-    hash: String,
-    senderSignature: String,
-    senderPublicKey: String,
+    footer: TransactionFooter,
 
     @Embedded
     var payload: TransferTransactionPayload
 
-) : UnconfirmedTransaction(header, hash, senderSignature, senderPublicKey) {
+) : UnconfirmedTransaction(header, footer) {
 
     companion object {
         fun of(message: TransferTransactionMessage): UnconfirmedTransferTransaction = UnconfirmedTransferTransaction(
             TransactionHeader(message.timestamp, message.fee, message.senderAddress),
-            message.hash,
-            message.senderSignature,
-            message.senderPublicKey,
+            TransactionFooter(message.hash, message.senderSignature, message.senderPublicKey),
             TransferTransactionPayload(message.amount, message.recipientAddress)
         )
 
         fun of(request: TransferTransactionRequest): UnconfirmedTransferTransaction = UnconfirmedTransferTransaction(
             TransactionHeader(request.timestamp!!, request.fee!!, request.senderAddress!!),
-            request.hash!!,
-            request.senderSignature!!,
-            request.senderPublicKey!!,
+            TransactionFooter(request.hash!!, request.senderSignature!!, request.senderPublicKey!!),
             TransferTransactionPayload(request.amount!!, request.recipientAddress!!)
         )
     }
-
-    override fun getPayload(): TransactionPayload = payload
 
     override fun toMessage(): TransferTransactionMessage = TransferTransactionMessage(
         header.timestamp,
         header.fee,
         header.senderAddress,
-        hash,
-        senderSignature,
-        senderPublicKey,
+        footer.hash,
+        footer.senderSignature,
+        footer.senderPublicKey,
         payload.amount,
         payload.recipientAddress
     )