+++ src/main/kotlin/io/openfuture/chain/core/service/transaction/DefaultTransferTransactionService.kt
@@ -1,7 +1,7 @@
 package io.openfuture.chain.core.service.transaction
 
-import io.openfuture.chain.core.component.TransactionCapacityChecker
 import io.openfuture.chain.core.annotation.BlockchainSynchronized
+import io.openfuture.chain.core.component.TransactionCapacityChecker
 import io.openfuture.chain.core.exception.NotFoundException
 import io.openfuture.chain.core.exception.ValidationException
 import io.openfuture.chain.core.exception.model.ExceptionType.INSUFFICIENT_BALANCE
@@ -12,7 +12,6 @@ import io.openfuture.chain.core.repository.TransferTransactionRepository
 import io.openfuture.chain.core.repository.UTransferTransactionRepository
 import io.openfuture.chain.core.service.TransferTransactionService
 import io.openfuture.chain.network.message.core.TransferTransactionMessage
-import io.openfuture.chain.network.service.NetworkApiService
 import io.openfuture.chain.rpc.domain.base.PageRequest
 import io.openfuture.chain.rpc.domain.transaction.request.TransferTransactionRequest
 import org.slf4j.LoggerFactory
@@ -24,18 +23,17 @@ import org.springframework.transaction.annotation.Transactional
 class DefaultTransferTransactionService(
     repository: TransferTransactionRepository,
     uRepository: UTransferTransactionRepository,
-    capacityChecker: TransactionCapacityChecker,
-    private val networkService: NetworkApiService
-) : BaseTransactionService<TransferTransaction, UnconfirmedTransferTransaction>(repository, uRepository, capacityChecker), TransferTransactionService {
-
-    @Transactional(readOnly = true)
-    override fun getByHash(hash: String): TransferTransaction = repository.findOneByHash(hash)
-        ?: throw NotFoundException("Transaction with hash $hash not found")
+    capacityChecker: TransactionCapacityChecker
+) : ExternalTransactionService<TransferTransaction, UnconfirmedTransferTransaction>(repository, uRepository, capacityChecker), TransferTransactionService {
 
     companion object {
         val log = LoggerFactory.getLogger(DefaultTransferTransactionService::class.java)
     }
 
+
+    @Transactional(readOnly = true)
+    override fun getByHash(hash: String): TransferTransaction = repository.findOneByFooterHash(hash)
+        ?: throw NotFoundException("Transaction with hash $hash not found")
 
     @Transactional(readOnly = true)
     override fun getAll(request: PageRequest): Page<TransferTransaction> = repository.findAll(request)
@@ -44,7 +42,7 @@ class DefaultTransferTransactionService(
     override fun getAllUnconfirmed(): MutableList<UnconfirmedTransferTransaction> = unconfirmedRepository.findAllByOrderByHeaderFeeDesc()
 
     @Transactional(readOnly = true)
-    override fun getUnconfirmedByHash(hash: String): UnconfirmedTransferTransaction = unconfirmedRepository.findOneByHash(hash)
+    override fun getUnconfirmedByHash(hash: String): UnconfirmedTransferTransaction = unconfirmedRepository.findOneByFooterHash(hash)
         ?: throw NotFoundException("Transaction with hash $hash not found")
 
     @Transactional(readOnly = true)
@@ -57,41 +55,23 @@ class DefaultTransferTransactionService(
 
     @Transactional
     override fun add(message: TransferTransactionMessage): UnconfirmedTransferTransaction {
-        val utx = UnconfirmedTransferTransaction.of(message)
-
-        if (isExists(utx.hash)) {
-            return utx
-        }
-
-        validate(utx)
-        val savedUtx = this.save(utx)
-        networkService.broadcast(message)
-        return savedUtx
+        return super.add(UnconfirmedTransferTransaction.of(message))
     }
 
-    @BlockchainSynchronized(throwable = true)
+    @BlockchainSynchronized
     @Transactional
     override fun add(request: TransferTransactionRequest): UnconfirmedTransferTransaction {
-        val utx = UnconfirmedTransferTransaction.of(request)
-
-        if (isExists(utx.hash)) {
-            return utx
-        }
-
-        validate(utx)
-        val savedUtx = this.save(utx)
-        networkService.broadcast(savedUtx.toMessage())
-        return savedUtx
+        return super.add(UnconfirmedTransferTransaction.of(request))
     }
 
     @Transactional
     override fun toBlock(message: TransferTransactionMessage, block: MainBlock): TransferTransaction {
-        val tx = repository.findOneByHash(message.hash)
+        val tx = repository.findOneByFooterHash(message.hash)
         if (null != tx) {
             return tx
         }
 
-        val utx = unconfirmedRepository.findOneByHash(message.hash)
+        val utx = unconfirmedRepository.findOneByFooterHash(message.hash)
         if (null != utx) {
             return confirm(utx, TransferTransaction.of(utx, block))
         }
@@ -101,26 +81,28 @@ class DefaultTransferTransactionService(
 
     @Transactional
     override fun verify(message: TransferTransactionMessage): Boolean {
-        return try {
+        try {
             validate(UnconfirmedTransferTransaction.of(message))
-            true
+            return true
         } catch (e: ValidationException) {
             log.warn(e.message)
-            false
+            return false
         }
     }
 
+    @Transactional
     override fun save(tx: TransferTransaction): TransferTransaction {
         updateTransferBalance(tx.header.senderAddress, tx.payload.recipientAddress, tx.payload.amount)
         return super.save(tx)
     }
 
-    private fun validate(utx: UnconfirmedTransferTransaction) {
+    @Transactional
+    override fun validate(utx: UnconfirmedTransferTransaction) {
         if (!isValidBalance(utx.header.senderAddress, utx.payload.amount, utx.header.fee)) {
             throw ValidationException("Insufficient balance", INSUFFICIENT_BALANCE)
         }
 
-        super.validateBase(utx)
+        super.validateExternal(utx.header, utx.payload, utx.footer)
     }
 
     private fun updateTransferBalance(from: String, to: String, amount: Long) {