+++ src/main/kotlin/io/openfuture/chain/core/service/transaction/DefaultVoteTransactionService.kt
@@ -1,8 +1,8 @@
 package io.openfuture.chain.core.service.transaction
 
 import io.openfuture.chain.consensus.property.ConsensusProperties
-import io.openfuture.chain.core.component.TransactionCapacityChecker
 import io.openfuture.chain.core.annotation.BlockchainSynchronized
+import io.openfuture.chain.core.component.TransactionCapacityChecker
 import io.openfuture.chain.core.exception.NotFoundException
 import io.openfuture.chain.core.exception.ValidationException
 import io.openfuture.chain.core.exception.model.ExceptionType.*
@@ -15,7 +15,6 @@ import io.openfuture.chain.core.repository.VoteTransactionRepository
 import io.openfuture.chain.core.service.DelegateService
 import io.openfuture.chain.core.service.VoteTransactionService
 import io.openfuture.chain.network.message.core.VoteTransactionMessage
-import io.openfuture.chain.network.service.NetworkApiService
 import io.openfuture.chain.rpc.domain.transaction.request.VoteTransactionRequest
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
@@ -27,13 +26,8 @@ internal class DefaultVoteTransactionService(
     uRepository: UVoteTransactionRepository,
     capacityChecker: TransactionCapacityChecker,
     private val delegateService: DelegateService,
-    private val consensusProperties: ConsensusProperties,
-    private val networkService: NetworkApiService
-) : BaseTransactionService<VoteTransaction, UnconfirmedVoteTransaction>(repository, uRepository, capacityChecker), VoteTransactionService {
-
-    @Transactional(readOnly = true)
-    override fun getByHash(hash: String): VoteTransaction = repository.findOneByHash(hash)
-        ?: throw NotFoundException("Transaction with hash $hash not found")
+    private val consensusProperties: ConsensusProperties
+) : ExternalTransactionService<VoteTransaction, UnconfirmedVoteTransaction>(repository, uRepository, capacityChecker), VoteTransactionService {
 
     companion object {
         val log = LoggerFactory.getLogger(DefaultVoteTransactionService::class.java)
@@ -41,49 +35,35 @@ internal class DefaultVoteTransactionService(
 
 
     @Transactional(readOnly = true)
+    override fun getByHash(hash: String): VoteTransaction = repository.findOneByFooterHash(hash)
+        ?: throw NotFoundException("Transaction with hash $hash not found")
+
+    @Transactional(readOnly = true)
     override fun getAllUnconfirmed(): MutableList<UnconfirmedVoteTransaction> = unconfirmedRepository.findAllByOrderByHeaderFeeDesc()
 
     @Transactional(readOnly = true)
-    override fun getUnconfirmedByHash(hash: String): UnconfirmedVoteTransaction = unconfirmedRepository.findOneByHash(hash)
+    override fun getUnconfirmedByHash(hash: String): UnconfirmedVoteTransaction = unconfirmedRepository.findOneByFooterHash(hash)
         ?: throw NotFoundException("Transaction with hash $hash not found")
 
     @Transactional
     override fun add(message: VoteTransactionMessage): UnconfirmedVoteTransaction {
-        val utx = UnconfirmedVoteTransaction.of(message)
-
-        if (isExists(utx.hash)) {
-            return utx
-        }
-
-        validate(utx)
-        val savedUtx = this.save(utx)
-        networkService.broadcast(message)
-        return savedUtx
+        return super.add(UnconfirmedVoteTransaction.of(message))
     }
 
     @BlockchainSynchronized
     @Transactional
     override fun add(request: VoteTransactionRequest): UnconfirmedVoteTransaction {
-        val utx = UnconfirmedVoteTransaction.of(request)
-        validate(utx)
-
-        if (isExists(utx.hash)) {
-            return utx
-        }
-
-        val savedUtx = this.save(utx)
-        networkService.broadcast(savedUtx.toMessage())
-        return savedUtx
+        return super.add(UnconfirmedVoteTransaction.of(request))
     }
 
     @Transactional
     override fun toBlock(message: VoteTransactionMessage, block: MainBlock): VoteTransaction {
-        val tx = repository.findOneByHash(message.hash)
+        val tx = repository.findOneByFooterHash(message.hash)
         if (null != tx) {
             return tx
         }
 
-        val utx = unconfirmedRepository.findOneByHash(message.hash)
+        val utx = unconfirmedRepository.findOneByFooterHash(message.hash)
         if (null != utx) {
             return confirm(utx, VoteTransaction.of(utx, block))
         }
@@ -93,22 +73,24 @@ internal class DefaultVoteTransactionService(
 
     @Transactional
     override fun verify(message: VoteTransactionMessage): Boolean {
-        return try {
+        try {
             validate(UnconfirmedVoteTransaction.of(message))
-            true
+            return true
         } catch (e: ValidationException) {
             log.warn(e.message)
-            false
+            return false
         }
     }
 
+    @Transactional
     override fun save(tx: VoteTransaction): VoteTransaction {
         val type = tx.payload.getVoteType()
         updateWalletVotes(tx.header.senderAddress, tx.payload.delegateKey, type)
         return super.save(tx)
     }
 
-    private fun validate(utx: UnconfirmedVoteTransaction) {
+    @Transactional
+    override fun validate(utx: UnconfirmedVoteTransaction) {
         if (!isExistsDelegate(utx.payload.delegateKey)) {
             throw ValidationException("Incorrect delegate key", INCORRECT_DELEGATE_KEY)
         }
@@ -129,7 +111,7 @@ internal class DefaultVoteTransactionService(
             throw ValidationException("Insufficient balance", INSUFFICIENT_BALANCE)
         }
 
-        super.validateBase(utx)
+        super.validateExternal(utx.header, utx.payload, utx.footer)
     }
 
     private fun updateWalletVotes(senderAddress: String, delegateKey: String, type: VoteType) {